[Bronze III] Title: 알고리즘 수업 - 알고리즘의 수행 시간 4, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Bronze/24265. 알고리즘 수업 － 알고리즘의 수행 시간 4/알고리즘 수업 － 알고리즘의 수행 시간 4.cc
@@ -2,9 +2,9 @@
 using namespace std;
 
 int main() {
-    int n;
+    unsigned long long n;
     cin >> n;
-    cout << (unsigned long long)n * (n - 1) / 2 << '\n';
+    cout << n * (n - 1) / 2 << '\n';
     cout << "2\n";
     return 0;
 }
